Fixed the _project!
Moved everything back into the _project/ folder, as it now works, and
converted to markdown. Still need to edit for succinctness.

Swapped Shashank's headshot to get rid of the black border

To run local: `make serve` inspecting the files locally is insufficient

Removed venkat from chargerplacement and fleetanalysis as requested:
https://eeg-group.slack.com/archives/CC4FWTM3L/p1620162281079200?thread_ts=1620158700.077400&cid=CC4FWTM3L